chore(test): refactor and add more tests
