+++ src/Worksheet.js
@@ -35,12 +35,22 @@ class Worksheet extends NamedItemContainer {
     return this._log;
   }
 
+  async getData() {
+    try {
+      const client = await this._oneDrive.getClient();
+      const result = await client.get(this._uri);
+      return result.value;
+    } catch (e) {
+      this.log.error(e);
+      throw new StatusCodeError(e.msg, 500);
+    }
+  }
+
   async getTableNames() {
     try {
       const client = await this._oneDrive.getClient();
       this.log.debug(`get table names from ${this._uri}/tables`);
       const result = await client.get(`${this._uri}/tables`);
-      console.log(result);
       return result.value.map((v) => v.name);
     } catch (e) {
       this.log.error(e);
